improved coding agent context, seperate out agent coding dialog component
